mv project config parser to settings
