+++ app/core/app.py
@@ -1,11 +1,14 @@
 from fastapi import FastAPI
-from app.utils.swagger_params import VERSION, DESCRIPTION, CONTACTS
 
+from app.core.settings import Settings
+
+settings = Settings()
+config = settings.ProjectConfigParser()
 app = FastAPI(
-    title="Joke API",
-    description=DESCRIPTION,
-    version=VERSION,
-    contact=CONTACTS,
+    title=settings.app_name,
+    description=config.get_project_description(),
+    version=config.get_project_version(),
+    contact=config.get_project_contacts(),
 )
